Harden security: lock down H2 console, fail-fast JWT secret, fix exception leakage and registration race; migrate currency fields from double to BigDecimal

+++ src/main/java/com/expensetracker/expense/dto/ExpenseRequest.java
@@ -2,9 +2,10 @@ package com.expensetracker.expense.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 
 public record ExpenseRequest(
         @NotBlank(message = "title is required") String title,
-        @Positive(message = "amount must be greater than 0") double amount,
+        @Positive(message = "amount must be greater than 0") BigDecimal amount,
         @NotBlank(message = "category is required") String category
 ) {}
