+++ src/main/java/com/expensetracker/savings/dto/SavingRequest.java
@@ -1,7 +1,8 @@
 package com.expensetracker.savings.dto;
 
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 
 public record SavingRequest(
-        @Positive(message = "amount must be greater than 0") double amount
+        @Positive(message = "amount must be greater than 0") BigDecimal amount
 ) {}
