+++ src/main/java/com/expensetracker/savings/dto/SavingsSummaryResponse.java
@@ -1,5 +1,6 @@
 package com.expensetracker.savings.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record SavingsSummaryResponse(double totalSaved, List<SavingResponse> savings) {}
+public record SavingsSummaryResponse(BigDecimal totalSaved, List<SavingResponse> savings) {}
